+++ lib/slack/web/api/patches.rb
@@ -1,1 +1,0 @@
-require 'slack/web/api/patches/chat'