+++ spec/slack/web/api/endpoints/chat_spec.rb
@@ -12,5 +12,22 @@ RSpec.describe Slack::Web::Api::Endpoints::Chat do
       )
       client.chat_postMessage(channel: 'channel', text: 'text', attachments: [])
     end
+    context 'text and attachment arguments' do
+      it 'requires text or attachments' do
+        expect { client.chat_postMessage(channel: 'channel') }.to raise_error ArgumentError, /Required arguments :text or :attachments missing/
+      end
+      it 'only text' do
+        expect(client).to receive(:post).with('chat.postMessage', hash_including(text: 'text'))
+        expect { client.chat_postMessage(channel: 'channel', text: 'text') }.to_not raise_error
+      end
+      it 'only attachments' do
+        expect(client).to receive(:post).with('chat.postMessage', hash_including(attachments: '[]'))
+        expect { client.chat_postMessage(channel: 'channel', attachments: []) }.to_not raise_error
+      end
+      it 'both text and attachments' do
+        expect(client).to receive(:post).with('chat.postMessage', hash_including(text: 'text', attachments: '[]'))
+        expect { client.chat_postMessage(channel: 'channel', text: 'text', attachments: []) }.to_not raise_error
+      end
+    end
   end
 end